fix: include the migration, library page
